(router): fix drawer navigator navigation (#25985)
# Why

Fix: #25981

# How

To be honest, I'm not 100% sure why its broken in the current state. It
fires a navigation events on the shared Stack with what seems like the
correct params.

But we also don't need to track the last common Stack/Tab
lastCommonNavigator. Pre-#24457, we used to fire navigation events on
the root navigator, so #24457 tried to not deviate from this behaviour
by only changing the behaviour Stacks&Tabs doing replace actions. #24600
cleaned up the logic of #24457 and unified the logic for all navigation
types, but still only tracked the last common Stack/Tabs. I don't know
why this breaks Drawer, it might be an internal implementation detail,
but we can simply fire the event on the last common navigator - no
matter what type it is.

# Test Plan

Add a smoke test for `<Drawer />`. I don't believe it was being
previously tested.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>