[router]: flush all test timers after each navigation to prevent false positives (#27981)
# Why

Expo Router's test suite is vulnerable to a false positive scenario,
where we may assert that a screen/pathname/etc is correct, but on the
next tick the app will rerender with a different state.

An example scenario is #27910. If we test this issue with `renderRouter`
it will pass but this is false positive. It will actually fail on a
device. This issue is caused by React Navigation having async state
updates on component mount and unmount. As `expo-router/testing-library`
uses fake timers (due to animations on the navigators), these updates
are "frozen" and will not execute. For the 1st frozen tick, the state
will be correct, but on the next tick and unmounting component may
update the state.

There maybe a better way to fix this, but for now I'm simply running all
pending timers after each navigation. This is a short term fix until we
can remove fake timers altogether.

# How

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>