chore(metro-runtime): stop transforming to cjs (#30300)
# Why

As a quick hack for SSR we transformed expo/metro-runtime to CJS so we
could import some error utils in the CLI without transforming via Metro.
This caused all of react-native-web to be imported on every request
(which will still happen in expo-router for the same reason). To slowly
decouple the processes, I'm copying over the symbolication code to the
CLI and removing the build dir from the expo/metro-runtime package.

# Test Plan

- apps/router-e2e -> yarn start -> web -> Add syntax errors and missing
imports. Both should print in the CLI as expected.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>